# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "DevicePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DeviceListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "DevicePage",
        "type": "object"
      },
...
      "DevicePositionsUpdate": {
        "description": "Расположение нескольких узлов сразу — результат автоматической\nраскладки схемы. Ограничение сверху не столько от злого умысла, сколько\nот опечатки в клиенте: схема на площадке больше пяти тысяч устройств не\nрисуется в любом случае.",
        "properties": {
          "positions": {
            "items": {
              "$ref": "#/components/schemas/DevicePosition"
            },
            "maxItems": 5000,
            "title": "Positions",
            "type": "array"
          }
        },
        "title": "DevicePositionsUpdate",
        "type": "object"
      },
      "DeviceTagsUpdate": {
        "properties": {
          "tag_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Tag Ids",
            "type": "array"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "tag_ids"
        ],
        "title": "DeviceTagsUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Network Documentation API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/devices": {
      "get": {
        "description": "Список устройств — страницами и без портов.\n\nПорты в списке не нужны и составляют почти весь его вес: на тысяче\nустройств по 24 порта это двадцать четыре тысячи вложенных объектов.\nЗдесь вместо них два числа — всего и занято, — а сами порты подтягивает\nраскрытая карточка (`GET /devices/{id}/interfaces`).\n\nОтбор и поиск считает база, а не интерфейс: фильтровать тысячу записей на\nстороне клиента можно, только сначала их туда привезя.",
        "operationId": "list_devices_devices_get",
        "parameters": [
          {
            "in": "query",
            "name": "tag_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag Id"
            }
          },
          {
            "in": "query",
            "name": "template_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Template Id"
            }
          },
          {
            "in": "query",
            "name": "device_type_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Type Id"
            }
          },
          {
            "in": "query",
            "name": "topology_group_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topology Group Id"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "in": "query",
            "name": "management_ip",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Management Ip"
            }
          },
          {
            "in": "query",
            "name": "mac",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mac"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "code",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "desc",
            "required": false,
            "schema": {
              "default": false,
              "title": "Desc",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Site-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Site-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevicePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Devices",
        "tags": [
          "devices"
        ]
      },
      "post": {
        "operationId": "create_device_devices_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Site-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Site-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Device",
        "tags": [
          "devices"
        ]
      }
    },
    "/devices/positions": {
      "patch": {
        "description": "Расположение сразу всех узлов схемы — одним запросом.\n\nАвтоматическая раскладка двигает не одну железку, а всю схему: полторы\nсотни отдельных PATCH-ов на одно нажатие кнопки — это полторы сотни\nтранзакций и столько же сериализаций устройства в ответ, из которых\nклиенту не нужна ни одна.\n\nОбъявлено выше `/{device_id}`: иначе «positions» попадёт в этот путь как\nидентификатор. Про чужие идентификаторы запрос молчит — раскладка могла\nсчитаться по схеме, из которой устройство успели удалить, и ронять на\nэтом сохранение остальных незачем.",
        "operationId": "update_device_positions_devices_positions_patch",
        "parameters": [
          {
            "in": "header",
            "name": "X-Site-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Site-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevicePositionsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Device Positions",
        "tags": [
          "devices"
        ]
      }
    },
...
    "/devices/{device_id}/interfaces": {
      "get": {
        "operationId": "list_interfaces_devices__device_id__interfaces_get",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Site-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Site-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InterfaceOut"
                  },
                  "title": "Response List Interfaces Devices  Device Id  Interfaces Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Interfaces",
        "tags": [
          "devices"
        ]
      },
      "post": {
        "description": "Добавить порт конкретному устройству.\n\nРазрешено только моделям с изменяемым составом портов — например ПК,\nв который доставили сетевую карту.",
        "operationId": "add_interface_devices__device_id__interfaces_post",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Site-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Site-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterfaceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterfaceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Add Interface",
        "tags": [
          "devices"
        ]
      }
    },
...
    "/devices/{device_id}/tags": {
      "put": {
        "operationId": "set_device_tags_devices__device_id__tags_put",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Site-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Site-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceTagsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Set Device Tags",
        "tags": [
          "devices"
        ]
      }
    },
...
  }
}